Use package module to install distro packages
Consolidate distro package install tasks into a
single task using the package module and pass
the package list into the name instead of using
a with_items loop.

The minimum Ansible version is raised to 2.2 due to a
known bug [1] in Ansible's apt module which does not
update the cache properly if the cache update and the
install are combined in a single task.

[1] https://github.com/ansible/ansible-modules-core/issues/1497

Also:
- split install tasks into pre-install, install and
  post-install tasks and tag them appropriately.

Change-Id: I0c5cc16ba1d82b7fd169156d1890c86614210cad