Only replace python when re-initializing the venv
We reinitialize the venv to ensure that the right version of
python is in the venv, but we do not want virtualenv to also
replace pip, setuptools and wheel so we tell it not to. If we
don't do this then virtualenv will install the latest available
version, which is not what we want.

Change-Id: Icc757ac36b298e46c0f257cd8304ac8553b85ee3
Partial-Bug: #1764470
(cherry picked from commit 0dae54128d45e7be3f1a823cdd8080e7ff536c78)