Be clear about python3: Remove its support
We have never really tested python3 at large scale, and more and
more bugs are filed in stable branches due to dual stack.

We should stick to python2 in the stable branches, which is
what was tested in our gates.

This also removes ubuntu 14.04 which should have been removed
long ago.

Change-Id: I37636e5cc1fdd149d401c03a53f3fe60f356c06f